requirements.txt
ultralytics
geopandas
folium
fastapi
uvicorn
python-multipart
reportlab
celery
redis
gevent
psycopg2-binary
sqlalchemy
geoalchemy2
python-jose
python-multipart
pillow